Moving the test utils into the package
